moved inspect_string_with_exclusion_dict into separate file
